refactor: Split notebook_edit.rs (1378 LOC) into validation + cell_ops modules
Decomposes the largest tools crate file into focused modules:
- notebook_edit/mod.rs: NotebookEditTool struct, types, Tool impl, execute (~428 LOC)
- notebook_edit/validation.rs: NotebookEditError, ValidationResult, validate_input (~136 LOC)
- notebook_edit/cell_ops.rs: resolve_cell_index, generate_cell_id, format_json, helpers (~143 LOC)
- notebook_edit/tests.rs: All 23 tests preserved and passing (~698 LOC)

No behavioral changes. lib.rs re-export unchanged (directory module resolves automatically).
