+++ crates/tools/src/notebook_edit/tests.rs
@@ -1,0 +1,698 @@
+//! Tests for NotebookEdit tool
+
+use super::cell_ops::*;
+use super::validation::*;
+use super::*;
+use crate::Tool;
+use futures::StreamExt;
+use std::io::Write;
+use tempfile::NamedTempFile;
+
+/// Helper to create a test notebook
+fn create_test_notebook(include_ids: bool, nbformat_minor: u8) -> serde_json::Value {
+    let cells = if include_ids {
+        serde_json::json!([
+            {
+                "cell_type": "code",
+                "id": "cell-1",
+                "source": "print('first')",
+                "metadata": {},
+                "execution_count": 1,
+                "outputs": [{"output_type": "stream", "text": "first\n"}]
+            },
+            {
+                "cell_type": "markdown",
+                "id": "cell-2",
+                "source": "# Title",
+                "metadata": {}
+            },
+            {
+                "cell_type": "code",
+                "id": "cell-3",
+                "source": "print('last')",
+                "metadata": {},
+                "execution_count": null,
+                "outputs": []
+            }
+        ])
+    } else {
+        serde_json::json!([
+            {
+                "cell_type": "code",
+                "source": "print('first')",
+                "metadata": {},
+                "execution_count": 1,
+                "outputs": [{"output_type": "stream", "text": "first\n"}]
+            },
+            {
+                "cell_type": "markdown",
+                "source": "# Title",
+                "metadata": {}
+            },
+            {
+                "cell_type": "code",
+                "source": "print('last')",
+                "metadata": {},
+                "execution_count": null,
+                "outputs": []
+            }
+        ])
+    };
+
+    serde_json::json!({
+        "cells": cells,
+        "metadata": {
+            "language_info": {
+                "name": "python"
+            }
+        },
+        "nbformat": 4,
+        "nbformat_minor": nbformat_minor
+    })
+}
+
+#[tokio::test]
+async fn test_parse_cell_index() {
+    assert_eq!(parse_cell_index("0"), Some(0));
+    assert_eq!(parse_cell_index("42"), Some(42));
+    assert_eq!(parse_cell_index("abc"), None);
+    assert_eq!(parse_cell_index("12a"), None);
+    assert_eq!(parse_cell_index(""), None);
+}
+
+#[tokio::test]
+async fn test_resolve_cell_index_numeric() {
+    let notebook = create_test_notebook(true, 5);
+    let cells = notebook.get("cells").unwrap().as_array().unwrap();
+
+    let (idx, is_numeric) = resolve_cell_index(cells, Some("0"), EditMode::Replace).unwrap();
+    assert_eq!(idx, 0);
+    assert!(is_numeric);
+
+    let (idx, is_numeric) = resolve_cell_index(cells, Some("2"), EditMode::Replace).unwrap();
+    assert_eq!(idx, 2);
+    assert!(is_numeric);
+}
+
+#[tokio::test]
+async fn test_resolve_cell_index_string_id() {
+    let notebook = create_test_notebook(true, 5);
+    let cells = notebook.get("cells").unwrap().as_array().unwrap();
+
+    let (idx, is_numeric) = resolve_cell_index(cells, Some("cell-1"), EditMode::Replace).unwrap();
+    assert_eq!(idx, 0);
+    assert!(!is_numeric);
+
+    let (idx, is_numeric) = resolve_cell_index(cells, Some("cell-3"), EditMode::Replace).unwrap();
+    assert_eq!(idx, 2);
+    assert!(!is_numeric);
+}
+
+#[tokio::test]
+async fn test_resolve_cell_index_not_found() {
+    let notebook = create_test_notebook(true, 5);
+    let cells = notebook.get("cells").unwrap().as_array().unwrap();
+
+    let result = resolve_cell_index(cells, Some("nonexistent"), EditMode::Replace);
+    assert!(result.is_err());
+    assert!(result.unwrap_err().to_string().contains("not found"));
+}
+
+#[tokio::test]
+async fn test_generate_cell_id() {
+    let id = generate_cell_id();
+    assert_eq!(id.len(), 13);
+    assert!(id.chars().all(|c| c.is_ascii_alphanumeric()));
+}
+
+#[tokio::test]
+async fn test_extract_language() {
+    let notebook = serde_json::json!({
+        "metadata": {
+            "language_info": {
+                "name": "julia"
+            }
+        }
+    });
+    assert_eq!(extract_language(&notebook), "julia");
+
+    let notebook_no_lang = serde_json::json!({"metadata": {}});
+    assert_eq!(extract_language(&notebook_no_lang), "python");
+}
+
+#[tokio::test]
+async fn test_supports_cell_ids() {
+    let notebook_45 = serde_json::json!({
+        "nbformat": 4,
+        "nbformat_minor": 5
+    });
+    assert!(supports_cell_ids(&notebook_45));
+
+    let notebook_44 = serde_json::json!({
+        "nbformat": 4,
+        "nbformat_minor": 4
+    });
+    assert!(!supports_cell_ids(&notebook_44));
+
+    let notebook_50 = serde_json::json!({
+        "nbformat": 5,
+        "nbformat_minor": 0
+    });
+    assert!(supports_cell_ids(&notebook_50));
+}
+
+#[tokio::test]
+async fn test_notebook_edit_replace_by_index() {
+    let notebook_json = create_test_notebook(true, 5);
+
+    let mut temp_file = NamedTempFile::with_suffix(".ipynb").unwrap();
+    write!(
+        temp_file,
+        "{}",
+        serde_json::to_string(&notebook_json).unwrap()
+    )
+    .unwrap();
+    temp_file.flush().unwrap();
+
+    let tool = NotebookEditTool;
+    let params = NotebookEditParams {
+        notebook_path: temp_file.path().to_str().unwrap().to_string(),
+        new_source: "print('modified')".to_string(),
+        cell_id: Some("0".to_string()),
+        cell_type: None,
+        edit_mode: EditMode::Replace,
+    };
+    let ctx = crate::ToolContext::default();
+
+    let stream = tool.execute(params, &ctx).await.unwrap();
+    let events: Vec<_> = stream.collect().await;
+
+    let result = events
+        .iter()
+        .find_map(|e| match e {
+            crate::ToolEvent::Result(output) => Some(output),
+            _ => None,
+        })
+        .unwrap();
+
+    assert_eq!(result.edit_mode, "replace");
+    assert_eq!(result.language, "python");
+    if let Some(ref err) = result.error {
+        eprintln!("Error: {}", err);
+    }
+    assert!(result.error.is_none());
+
+    // Verify notebook was modified
+    let content = tokio::fs::read_to_string(temp_file.path()).await.unwrap();
+    assert!(content.contains("print('modified')"));
+    assert!(!content.contains("print('first')"));
+
+    // Verify execution state was reset
+    let notebook: serde_json::Value = serde_json::from_str(&content).unwrap();
+    let cell = &notebook["cells"][0];
+    assert_eq!(cell["execution_count"], serde_json::Value::Null);
+    assert_eq!(cell["outputs"], serde_json::Value::Array(vec![]));
+}
+
+#[tokio::test]
+async fn test_notebook_edit_replace_by_string_id() {
+    let notebook_json = create_test_notebook(true, 5);
+
+    let mut temp_file = NamedTempFile::with_suffix(".ipynb").unwrap();
+    write!(
+        temp_file,
+        "{}",
+        serde_json::to_string(&notebook_json).unwrap()
+    )
+    .unwrap();
+    temp_file.flush().unwrap();
+
+    let tool = NotebookEditTool;
+    let params = NotebookEditParams {
+        notebook_path: temp_file.path().to_str().unwrap().to_string(),
+        new_source: "print('updated')".to_string(),
+        cell_id: Some("cell-3".to_string()),
+        cell_type: None,
+        edit_mode: EditMode::Replace,
+    };
+    let ctx = crate::ToolContext::default();
+
+    let stream = tool.execute(params, &ctx).await.unwrap();
+    let events: Vec<_> = stream.collect().await;
+
+    let result = events
+        .iter()
+        .find_map(|e| match e {
+            crate::ToolEvent::Result(output) => Some(output),
+            _ => None,
+        })
+        .unwrap();
+
+    assert_eq!(result.cell_id, Some("cell-3".to_string()));
+    assert!(result.error.is_none());
+
+    let content = tokio::fs::read_to_string(temp_file.path()).await.unwrap();
+    assert!(content.contains("print('updated')"));
+}
+
+#[tokio::test]
+async fn test_notebook_edit_insert_with_cell_id() {
+    let notebook_json = create_test_notebook(true, 5);
+
+    let mut temp_file = NamedTempFile::with_suffix(".ipynb").unwrap();
+    write!(
+        temp_file,
+        "{}",
+        serde_json::to_string(&notebook_json).unwrap()
+    )
+    .unwrap();
+    temp_file.flush().unwrap();
+
+    let tool = NotebookEditTool;
+    let params = NotebookEditParams {
+        notebook_path: temp_file.path().to_str().unwrap().to_string(),
+        new_source: "print('inserted')".to_string(),
+        cell_id: Some("cell-1".to_string()),
+        cell_type: Some(CellType::Code),
+        edit_mode: EditMode::Insert,
+    };
+    let ctx = crate::ToolContext::default();
+
+    let stream = tool.execute(params, &ctx).await.unwrap();
+    let events: Vec<_> = stream.collect().await;
+
+    let result = events
+        .iter()
+        .find_map(|e| match e {
+            crate::ToolEvent::Result(output) => Some(output),
+            _ => None,
+        })
+        .unwrap();
+
+    assert_eq!(result.edit_mode, "insert");
+    assert!(result.error.is_none());
+
+    let content = tokio::fs::read_to_string(temp_file.path()).await.unwrap();
+    let notebook: serde_json::Value = serde_json::from_str(&content).unwrap();
+    assert_eq!(notebook["cells"].as_array().unwrap().len(), 4);
+    assert!(content.contains("print('inserted')"));
+}
+
+#[tokio::test]
+async fn test_notebook_edit_insert_at_beginning() {
+    let notebook_json = create_test_notebook(true, 5);
+
+    let mut temp_file = NamedTempFile::with_suffix(".ipynb").unwrap();
+    write!(
+        temp_file,
+        "{}",
+        serde_json::to_string(&notebook_json).unwrap()
+    )
+    .unwrap();
+    temp_file.flush().unwrap();
+
+    let tool = NotebookEditTool;
+    let params = NotebookEditParams {
+        notebook_path: temp_file.path().to_str().unwrap().to_string(),
+        new_source: "# New Title".to_string(),
+        cell_id: None,
+        cell_type: Some(CellType::Markdown),
+        edit_mode: EditMode::Insert,
+    };
+    let ctx = crate::ToolContext::default();
+
+    let stream = tool.execute(params, &ctx).await.unwrap();
+    let events: Vec<_> = stream.collect().await;
+
+    let result = events
+        .iter()
+        .find_map(|e| match e {
+            crate::ToolEvent::Result(output) => Some(output),
+            _ => None,
+        })
+        .unwrap();
+
+    assert!(result.error.is_none());
+
+    let content = tokio::fs::read_to_string(temp_file.path()).await.unwrap();
+    let notebook: serde_json::Value = serde_json::from_str(&content).unwrap();
+    assert_eq!(notebook["cells"][0]["source"], "# New Title");
+    assert_eq!(notebook["cells"][0]["cell_type"], "markdown");
+}
+
+#[tokio::test]
+async fn test_notebook_edit_delete() {
+    let notebook_json = create_test_notebook(true, 5);
+
+    let mut temp_file = NamedTempFile::with_suffix(".ipynb").unwrap();
+    write!(
+        temp_file,
+        "{}",
+        serde_json::to_string(&notebook_json).unwrap()
+    )
+    .unwrap();
+    temp_file.flush().unwrap();
+
+    let tool = NotebookEditTool;
+    let params = NotebookEditParams {
+        notebook_path: temp_file.path().to_str().unwrap().to_string(),
+        new_source: "".to_string(),
+        cell_id: Some("1".to_string()),
+        cell_type: None,
+        edit_mode: EditMode::Delete,
+    };
+    let ctx = crate::ToolContext::default();
+
+    let stream = tool.execute(params, &ctx).await.unwrap();
+    let events: Vec<_> = stream.collect().await;
+
+    let result = events
+        .iter()
+        .find_map(|e| match e {
+            crate::ToolEvent::Result(output) => Some(output),
+            _ => None,
+        })
+        .unwrap();
+
+    assert_eq!(result.edit_mode, "delete");
+    assert!(result.error.is_none());
+
+    let content = tokio::fs::read_to_string(temp_file.path()).await.unwrap();
+    let notebook: serde_json::Value = serde_json::from_str(&content).unwrap();
+    assert_eq!(notebook["cells"].as_array().unwrap().len(), 2);
+    assert!(!content.contains("# Title"));
+}
+
+#[tokio::test]
+async fn test_validation_file_not_found() {
+    let params = NotebookEditParams {
+        notebook_path: "/nonexistent/path.ipynb".to_string(),
+        new_source: "test".to_string(),
+        cell_id: Some("0".to_string()),
+        cell_type: None,
+        edit_mode: EditMode::Replace,
+    };
+
+    let result = validate_input(&params).await;
+    assert!(!result.result);
+    assert_eq!(result.error_code, Some(1));
+}
+
+#[tokio::test]
+async fn test_validation_invalid_file_type() {
+    let mut temp_file = NamedTempFile::with_suffix(".txt").unwrap();
+    write!(temp_file, "not a notebook").unwrap();
+    temp_file.flush().unwrap();
+
+    let params = NotebookEditParams {
+        notebook_path: temp_file.path().to_str().unwrap().to_string(),
+        new_source: "test".to_string(),
+        cell_id: Some("0".to_string()),
+        cell_type: None,
+        edit_mode: EditMode::Replace,
+    };
+
+    let result = validate_input(&params).await;
+    assert!(!result.result);
+    assert_eq!(result.error_code, Some(2));
+}
+
+#[tokio::test]
+async fn test_validation_missing_cell_type_for_insert() {
+    let notebook_json = create_test_notebook(true, 5);
+    let mut temp_file = NamedTempFile::with_suffix(".ipynb").unwrap();
+    write!(
+        temp_file,
+        "{}",
+        serde_json::to_string(&notebook_json).unwrap()
+    )
+    .unwrap();
+    temp_file.flush().unwrap();
+
+    let params = NotebookEditParams {
+        notebook_path: temp_file.path().to_str().unwrap().to_string(),
+        new_source: "test".to_string(),
+        cell_id: None,
+        cell_type: None,
+        edit_mode: EditMode::Insert,
+    };
+
+    let result = validate_input(&params).await;
+    assert!(!result.result);
+    assert_eq!(result.error_code, Some(5));
+}
+
+#[tokio::test]
+async fn test_validation_invalid_json() {
+    let mut temp_file = NamedTempFile::with_suffix(".ipynb").unwrap();
+    write!(temp_file, "{{invalid json}}").unwrap();
+    temp_file.flush().unwrap();
+
+    let params = NotebookEditParams {
+        notebook_path: temp_file.path().to_str().unwrap().to_string(),
+        new_source: "test".to_string(),
+        cell_id: Some("0".to_string()),
+        cell_type: None,
+        edit_mode: EditMode::Replace,
+    };
+
+    let result = validate_input(&params).await;
+    assert!(!result.result);
+    assert_eq!(result.error_code, Some(6));
+}
+
+#[tokio::test]
+async fn test_validation_missing_cell_id_for_replace() {
+    let notebook_json = create_test_notebook(true, 5);
+    let mut temp_file = NamedTempFile::with_suffix(".ipynb").unwrap();
+    write!(
+        temp_file,
+        "{}",
+        serde_json::to_string(&notebook_json).unwrap()
+    )
+    .unwrap();
+    temp_file.flush().unwrap();
+
+    let params = NotebookEditParams {
+        notebook_path: temp_file.path().to_str().unwrap().to_string(),
+        new_source: "test".to_string(),
+        cell_id: None,
+        cell_type: None,
+        edit_mode: EditMode::Replace,
+    };
+
+    let result = validate_input(&params).await;
+    assert!(!result.result);
+    assert_eq!(result.error_code, Some(7));
+}
+
+#[tokio::test]
+async fn test_validation_cell_not_found_numeric() {
+    let notebook_json = create_test_notebook(true, 5);
+    let mut temp_file = NamedTempFile::with_suffix(".ipynb").unwrap();
+    write!(
+        temp_file,
+        "{}",
+        serde_json::to_string(&notebook_json).unwrap()
+    )
+    .unwrap();
+    temp_file.flush().unwrap();
+
+    let params = NotebookEditParams {
+        notebook_path: temp_file.path().to_str().unwrap().to_string(),
+        new_source: "test".to_string(),
+        cell_id: Some("99".to_string()),
+        cell_type: None,
+        edit_mode: EditMode::Replace,
+    };
+
+    let result = validate_input(&params).await;
+    assert!(!result.result);
+    assert_eq!(result.error_code, Some(8));
+}
+
+#[tokio::test]
+async fn test_validation_cell_not_found_string_id() {
+    let notebook_json = create_test_notebook(true, 5);
+    let mut temp_file = NamedTempFile::with_suffix(".ipynb").unwrap();
+    write!(
+        temp_file,
+        "{}",
+        serde_json::to_string(&notebook_json).unwrap()
+    )
+    .unwrap();
+    temp_file.flush().unwrap();
+
+    let params = NotebookEditParams {
+        notebook_path: temp_file.path().to_str().unwrap().to_string(),
+        new_source: "test".to_string(),
+        cell_id: Some("nonexistent-id".to_string()),
+        cell_type: None,
+        edit_mode: EditMode::Replace,
+    };
+
+    let result = validate_input(&params).await;
+    assert!(!result.result);
+    assert_eq!(result.error_code, Some(8));
+}
+
+#[tokio::test]
+async fn test_smart_edit_mode_detection() {
+    // Smart mode detection happens during execution when cell_index == cells.length
+    // Validation rejects out-of-bounds, so we test with a valid scenario
+    let notebook_json = create_test_notebook(true, 5);
+    let mut temp_file = NamedTempFile::with_suffix(".ipynb").unwrap();
+    write!(
+        temp_file,
+        "{}",
+        serde_json::to_string(&notebook_json).unwrap()
+    )
+    .unwrap();
+    temp_file.flush().unwrap();
+
+    // Test that out-of-bounds numeric index in Replace mode fails validation
+    let params_fail = NotebookEditParams {
+        notebook_path: temp_file.path().to_str().unwrap().to_string(),
+        new_source: "print('fail')".to_string(),
+        cell_id: Some("99".to_string()), // Way out of bounds
+        cell_type: None,
+        edit_mode: EditMode::Replace,
+    };
+
+    let result = validate_input(&params_fail).await;
+    assert!(!result.result);
+    assert_eq!(result.error_code, Some(8));
+
+    // Now test the actual smart edit mode: replace beyond last existing cell
+    // In TypeScript (line 974-978), when cell_index == cells.length in Replace mode,
+    // it converts to Insert. However, since validation would reject index 3 (cells.length),
+    // this conversion actually happens during execution, not after validation.
+    // The Rust implementation should mirror this: reject in validation for safety.
+    // Therefore, this test confirms validation correctly rejects out-of-bounds Replace.
+}
+
+#[tokio::test]
+async fn test_cell_id_generation_for_nbformat_45() {
+    let notebook_json = create_test_notebook(false, 5);
+    let mut temp_file = NamedTempFile::with_suffix(".ipynb").unwrap();
+    write!(
+        temp_file,
+        "{}",
+        serde_json::to_string(&notebook_json).unwrap()
+    )
+    .unwrap();
+    temp_file.flush().unwrap();
+
+    let tool = NotebookEditTool;
+    let params = NotebookEditParams {
+        notebook_path: temp_file.path().to_str().unwrap().to_string(),
+        new_source: "print('new cell')".to_string(),
+        cell_id: None,
+        cell_type: Some(CellType::Code),
+        edit_mode: EditMode::Insert,
+    };
+    let ctx = crate::ToolContext::default();
+
+    let stream = tool.execute(params, &ctx).await.unwrap();
+    let events: Vec<_> = stream.collect().await;
+
+    let result = events
+        .iter()
+        .find_map(|e| match e {
+            crate::ToolEvent::Result(output) => Some(output),
+            _ => None,
+        })
+        .unwrap();
+
+    assert!(result.error.is_none());
+
+    let content = tokio::fs::read_to_string(temp_file.path()).await.unwrap();
+    let notebook: serde_json::Value = serde_json::from_str(&content).unwrap();
+    let inserted_cell = &notebook["cells"][0];
+
+    // Should have generated an ID
+    assert!(inserted_cell.get("id").is_some());
+    let id = inserted_cell["id"].as_str().unwrap();
+    assert_eq!(id.len(), 13);
+}
+
+#[tokio::test]
+async fn test_no_cell_id_for_old_nbformat() {
+    let notebook_json = create_test_notebook(false, 4);
+    let mut temp_file = NamedTempFile::with_suffix(".ipynb").unwrap();
+    write!(
+        temp_file,
+        "{}",
+        serde_json::to_string(&notebook_json).unwrap()
+    )
+    .unwrap();
+    temp_file.flush().unwrap();
+
+    let tool = NotebookEditTool;
+    let params = NotebookEditParams {
+        notebook_path: temp_file.path().to_str().unwrap().to_string(),
+        new_source: "print('new cell')".to_string(),
+        cell_id: None,
+        cell_type: Some(CellType::Code),
+        edit_mode: EditMode::Insert,
+    };
+    let ctx = crate::ToolContext::default();
+
+    let stream = tool.execute(params, &ctx).await.unwrap();
+    let events: Vec<_> = stream.collect().await;
+
+    let result = events
+        .iter()
+        .find_map(|e| match e {
+            crate::ToolEvent::Result(output) => Some(output),
+            _ => None,
+        })
+        .unwrap();
+
+    assert!(result.error.is_none());
+
+    let content = tokio::fs::read_to_string(temp_file.path()).await.unwrap();
+    let notebook: serde_json::Value = serde_json::from_str(&content).unwrap();
+    let inserted_cell = &notebook["cells"][0];
+
+    // Should NOT have generated an ID
+    assert!(inserted_cell.get("id").is_none());
+}
+
+#[tokio::test]
+async fn test_json_single_space_formatting() {
+    let notebook_json = create_test_notebook(true, 5);
+    let mut temp_file = NamedTempFile::with_suffix(".ipynb").unwrap();
+    write!(
+        temp_file,
+        "{}",
+        serde_json::to_string(&notebook_json).unwrap()
+    )
+    .unwrap();
+    temp_file.flush().unwrap();
+
+    let tool = NotebookEditTool;
+    let params = NotebookEditParams {
+        notebook_path: temp_file.path().to_str().unwrap().to_string(),
+        new_source: "print('test')".to_string(),
+        cell_id: Some("0".to_string()),
+        cell_type: None,
+        edit_mode: EditMode::Replace,
+    };
+    let ctx = crate::ToolContext::default();
+
+    let stream = tool.execute(params, &ctx).await.unwrap();
+    let _events: Vec<_> = stream.collect().await;
+
+    let content = tokio::fs::read_to_string(temp_file.path()).await.unwrap();
+
+    // Check that formatting uses single space indentation
+    let lines: Vec<&str> = content.lines().collect();
+    let has_single_space_indent = lines
+        .iter()
+        .any(|line| line.starts_with(" \"") && !line.starts_with("  "));
+    assert!(
+        has_single_space_indent,
+        "Should have single-space indentation"
+    );
+}
